feat: add CLI, modules and hash solutions
