+++ src/cli/args.js
@@ -1,5 +1,11 @@
 const parseArgs = () => {
-    // Write your code here 
+  console.log(process.argv.slice(2).reduce((acc, item, index) => {
+    if (index % 2 === 0) {
+      return index === 0 ? item.replace('--', '') : `${acc}, ${item.replace('--', '')}`;
+    } else {
+      return `${acc} is ${item}`;
+    }
+  }, ''));
 };
 
 parseArgs();
