+++ src/cli/env.js
@@ -1,5 +1,10 @@
 const parseEnv = () => {
-    // Write your code here 
+  console.log(Object.entries(process.env).reduce((acc, [key, value]) => {
+    if (key.includes('RSS_')) {
+      return acc === '' ? `${key}=${value}` : `${acc}; ${key}=${value}`;
+    }
+    return acc;
+  }, ''));
 };
 
 parseEnv();
